@@ -2,7 +2,12 @@
   "name": "oveleon/contao-component-style-manager",
   "type": "contao-bundle",
   "description": "Style and CSS-Class Manager for Contao Open Source CMS",
-  "keywords": ["contao", "styles", "css", "manager"],
+  "keywords": [
+    "contao",
+    "styles",
+    "css",
+    "manager"
+  ],
   "homepage": "https://www.oveleon.de/stylemanager.html",
   "license": "MIT",
   "authors": [
@@ -18,10 +23,21 @@
     }
   ],
   "require": {
-    "php": "^8.0",
+    "php": "^8.1",
     "ext-json": "*",
     "ext-dom": "*",
-    "contao/core-bundle": "^4.13 || ^5.1"
+    "contao/core-bundle": "^4.13 || ^5.1",
+    "doctrine/dbal": "^3.3",
+    "psr/log": "^1.0 || 2.0 || ^3.0",
+    "symfony/config": "^5.4 || ^6.4 || ^7.0",
+    "symfony/console": "^5.4 || ^6.4 || ^7.0",
+    "symfony/dependency-injection": "^5.4 || ^6.4 || ^7.0",
+    "symfony/event-dispatcher": "^5.4 || ^6.4 || ^7.0",
+    "symfony/http-foundation": "^5.4 || ^6.4 || ^7.0",
+    "symfony/http-kernel": "^5.4 || ^6.4 || ^7.0",
+    "symfony/routing": "^5.4 || ^6.4 || ^7.0",
+    "symfony/translation-contracts": "^2.3 || ^3.0",
+    "twig/twig": "^3.8"
   },
   "require-dev": {
     "contao/manager-plugin": "^2.3.1"
@@ -49,7 +65,7 @@
   },
   "extra": {
     "branch-alias": {
-      "dev-main": "3.4.x-dev"
+      "dev-main": "3.6.x-dev"
     },
     "contao-manager-plugin": "Oveleon\\ContaoComponentStyleManager\\ContaoManager\\Plugin"
   },
