@@ -10,22 +10,20 @@ declare(strict_types=1);
 
 namespace Oveleon\ContaoComponentStyleManager\ContaoManager;
 
-use Contao\CoreBundle\ContaoCoreBundle;
 use Contao\CalendarBundle\ContaoCalendarBundle;
-use Contao\ManagerPlugin\Routing\RoutingPluginInterface;
-use Contao\NewsBundle\ContaoNewsBundle;
+use Contao\CoreBundle\ContaoCoreBundle;
 use Contao\ManagerPlugin\Bundle\BundlePluginInterface;
 use Contao\ManagerPlugin\Bundle\Config\BundleConfig;
 use Contao\ManagerPlugin\Bundle\Parser\ParserInterface;
+use Contao\ManagerPlugin\Routing\RoutingPluginInterface;
+use Contao\NewsBundle\ContaoNewsBundle;
 use Oveleon\ContaoComponentStyleManager\ContaoComponentStyleManager;
 use Symfony\Component\Config\Loader\LoaderResolverInterface;
 use Symfony\Component\HttpKernel\KernelInterface;
+use Symfony\Component\Routing\RouteCollection;
 
 class Plugin implements BundlePluginInterface, RoutingPluginInterface
 {
-    /**
-     * {@inheritdoc}
-     */
     public function getBundles(ParserInterface $parser): array
     {
         return [
@@ -35,14 +33,11 @@ class Plugin implements BundlePluginInterface, RoutingPluginInterface
         ];
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function getRouteCollection(LoaderResolverInterface $resolver, KernelInterface $kernel)
+    public function getRouteCollection(LoaderResolverInterface $resolver, KernelInterface $kernel): RouteCollection|null
     {
         return $resolver
-            ->resolve(__DIR__.'/../../config/routes.yaml')
-            ->load(__DIR__.'/../../config/routes.yaml')
-            ;
+            ->resolve('@ContaoComponentStyleManager/src/Controller')
+            ->load('@ContaoComponentStyleManager/src/Controller')
+        ;
     }
 }
